move function definitions to main file.
